core/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Deprecated==1.2.10
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-countries==6.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-import-export==2.4.0
django-select2==7.4.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.1
jdcal==1.4.1
MarkupPy==1.14
nexmo==2.5.2
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
twilio==6.50.1
Unipath==1.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
